+++ domain/src/main/java/com/music/musicStreamer/gateway/ImageGateway.java
@@ -7,6 +7,6 @@ import com.music.musicStreamer.entity.image.UploadImageEntity;
 public interface ImageGateway {
     ImageEntity save(UploadImageEntity uploadImageEntity);
     byte[] getByFileName(final String pathName);
-    Boolean deleteByMusicId(final int musicId);
+    Boolean deleteByMusicId(final Long musicId);
     ImageEntity getByMusicId(final int musicId);
 }
